pyproject.toml
[tool.poetry]
name = "auth-service"
version = "0.1.0"
description = ""
authors = ["yogson <yogson@gmail.com>"]
readme = "README.md"
packages = [{include = "auth_service"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.104.1"
python-jose = "^3.3.0"
passlib = "^1.7.4"
uvicorn = "0.24.0"
python-multipart = "^0.0.6"
slowapi = "^0.1.8"
bcrypt = "^4.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120